BackendOptiTAB/requirements.txt
antlr4-python3-runtime==4.11.0
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
django-rest-passwordreset==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
mpmath==1.3.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.1
sqlparse==0.5.3
sympy==1.14.0
tzdata==2025.2
playwright==1.45.0
